Move user classes into unified namespace
