+++ site/src/App/EventListener/OAuthRegistrationListener.php
@@ -7,8 +7,8 @@ use FOS\UserBundle\Event\FormEvent;
 use FOS\UserBundle\Event\GetResponseUserEvent;
 use FOS\UserBundle\FOSUserEvents;
 use FOS\UserBundle\Model\UserManagerInterface;
-use Librecores\ProjectRepoBundle\Security\Core\Exception\OAuthUserLinkingException;
-use Librecores\ProjectRepoBundle\Security\Core\User\LibreCoresUserProvider;
+use App\Security\Core\Exception\OAuthUserLinkingException;
+use App\Security\Core\User\LibreCoresUserProvider;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
