+++ site/src/App/Security/Core/Exception/InsufficientOAuthDataProvidedException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Librecores\ProjectRepoBundle\Security\Core\Exception;
+namespace App\Security\Core\Exception;
 
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
