+++ site/src/Librecores/ProjectRepoBundle/Resources/config/legacy_aliases.yml
@@ -9,9 +9,9 @@ services:
     librecores.form.type.purified_textarea: '@App\Form\Type\PurifiedTextareaType'
     librecores.form.transformer.html_purifier: '@Exercise\HTMLPurifierBundle\Form\HTMLPurifierTransformer'
     validator.user_org_name: '@App\Validator\Constraints\UserOrgNameValidator'
-    librecores_user_manager: '@Librecores\ProjectRepoBundle\Doctrine\UserManager'
-    librecores_user_provider: '@Librecores\ProjectRepoBundle\Security\Core\User\LibreCoresUserProvider'
+    librecores_user_manager: '@App\Security\UserManager'
+    librecores_user_provider: '@App\Security\Core\User\LibreCoresUserProvider'
     librecores.user.registration_event_listener: '@App\EventListener\OAuthRegistrationListener'
-    librecores_project_voter: '@Librecores\ProjectRepoBundle\Security\ProjectVoter'
+    librecores_project_voter: '@App\Security\ProjectVoter'
     librecores.process_creator: '@App\Util\ProcessCreator'
     librecores.repocrawlerfactory: '@App\RepoCrawler\RepoCrawlerFactory'
