+++ site/src/Librecores/ProjectRepoBundle/Resources/config/services.yml
@@ -41,24 +41,4 @@ services:
         calls:
             - [setContainer, ['@service_container']]
 
-    # User/auth management
-    Librecores\ProjectRepoBundle\Doctrine\UserManager:
-        arguments:
-            - '@fos_user.util.password_updater'
-            - '@fos_user.util.canonical_fields_updater'
-            - '@fos_user.object_manager'
-            - '%fos_user.model.user.class%'
-
-    Librecores\ProjectRepoBundle\Security\Core\User\LibreCoresUserProvider:
-        arguments:
-            - '@fos_user.user_manager'
-            - '@validator'
-            - '@session'
-
-    # Security
-    Librecores\ProjectRepoBundle\Security\ProjectVoter:
-        tags:
-            - { name: security.voter }
-        public: false
-
     App\RepoCrawler\RepoCrawlerFactory: ~
